+++ src/changelog/commands.py
@@ -4,7 +4,15 @@ from changelog.utils import initialize_changelog_file, update_section, get_new_r
 from changelog.exceptions import ChangelogDoesNotExistError
 
 
+def print_version(ctx, _, value):
+    from changelog._version import __version__ as v
+    if not value or ctx.resilient_parsing:
+        return
+    click.echo(v)
+    ctx.exit()
+
 @click.group()
+@click.option('-v', '--version', is_flag=True, callback=print_version, expose_value=False, is_eager=True)
 def cli():
     pass
 
@@ -65,11 +73,23 @@ def breaks(message):
 @click.option('--minor', 'release_type', flag_value='minor')
 @click.option('--major', 'release_type', flag_value='major')
 @click.option('--suggest', 'release_type', flag_value='suggest', default=True)
-def release(release_type):
+@click.option('--yes', 'auto_confirm', is_flag=True)
+def release(release_type, auto_confirm):
     try:
         new_version = get_new_release_version(release_type)
-        if click.confirm("Planning on releasing version {}. Proceed?".format(new_version)):
-            cut_release(release_type)
+        if auto_confirm:
+            cut_release()
+        else:
+            if click.confirm("Planning on releasing version {}. Proceed?".format(new_version)):
+                cut_release(release_type)
     except ChangelogDoesNotExistError:
         if click.confirm("No CHANGELOG.md Found, do you want to create one?"):
             initialize_changelog_file()
+
+@cli.command()
+def suggest():
+    try:
+        new_version = get_new_release_version('suggest')
+        click.echo(new_version)
+    except ChangelogDoesNotExistError:
+        pass