+++ tasks/lint/__init__.py
@@ -1,0 +1,12 @@
+"""
+Code Linting
+"""
+from invoke import task
+
+
+@task
+def lint(context):
+    """
+    Runs quality checks through linter
+    """
+    context.run('prospector')