+++ tasks/test/__init__.py
@@ -1,0 +1,13 @@
+"""
+Run Test Suite
+"""
+
+from invoke import task
+
+
+@task
+def test(context):
+    """
+    Runs Test Suite
+    """
+    context.run('coverage run -m unittest discover')