+++ tests/unit/test_utils.py
@@ -1,6 +1,6 @@
 import unittest
 
-from changelog.utils import bump_version
+from changelog.utils import bump_version, crunch_lines
 
 class UtilsTestCase(unittest.TestCase):
 
@@ -13,3 +13,16 @@ class UtilsTestCase(unittest.TestCase):
     def test_bump_version_patch(self):
         self.assertEqual(bump_version('0.1.0', 'patch'), '0.1.1')
 
+    def test_crunch_lines(self):
+        document = [
+            "this\n",
+            "\n",
+            "\n",
+            "\n",
+            "\n",
+            "that\n"
+        ]
+        self.assertEqual(crunch_lines(document), ['this\n', '\n', '\n', 'that\n'])
+
+    def test_get_release_suggestion_patch(self):
+        pass